ExtractorHTML: Treat 'cite' attribute as navlink instead of embed
The cite attribute is used to identify the source document of a blockquote. But ExtractorHTML was treating it as an embed which can cause out of scope pages to be collected incorrectly. Browsers don't use it currently so there mightbe an argument for ignoring it entirely but let's at least not treat it as an embed.